# Codigo/supportplatform-service/src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ValidationType": {
				"enum": [
					"DEFAULT",
					"CFF"
				],
				"type": "string"
			},
			"ConsistencyRuleResponseDTO": {
				"description": "ConsistencyRuleResponseDTO - Could be used to structure the response object when sending ConsistencyRules to the client.",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"evaluationMethodId": {
						"type": "number",
						"format": "double"
					},
					"sprintId": {
						"type": "number",
						"format": "double"
					},
					"standardizedIssueId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"filePath": {
						"type": "string"
					},
					"validationType": {
						"$ref": "#/components/schemas/ValidationType"
					}
				},
				"required": [
					"id",
					"evaluationMethodId",
					"sprintId",
					"standardizedIssueId",
					"description",
					"filePath",
					"validationType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConsistencyRuleCreateDTO": {
				"description": "ConsistencyRuleCreateDTO - Used for creating a new ConsistencyRule.",
				"properties": {
					"evaluationMethodId": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"sprintId": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"standardizedIssueId": {
						"type": "integer",
						"format": "int32",
						"nullable": true,
						"minimum": 1
					},
					"description": {
						"type": "string",
						"nullable": true,
						"minLength": 1,
						"maxLength": 255
					},
					"filePath": {
						"type": "string",
						"minLength": 1,
						"maxLength": 10000
					},
					"validationType": {
						"$ref": "#/components/schemas/ValidationType"
					}
				},
				"required": [
					"evaluationMethodId",
					"sprintId",
					"filePath",
					"validationType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConsistencyRuleUpdateDTO": {
				"description": "ConsistencyRuleUpdateDTO - Used for updating an existing ConsistencyRule.",
				"properties": {
					"evaluationMethodId": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"sprintId": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"standardizedIssueId": {
						"type": "integer",
						"format": "int32",
						"nullable": true,
						"minimum": 1
					},
					"description": {
						"type": "string",
						"nullable": true,
						"minLength": 1,
						"maxLength": 255
					},
					"filePath": {
						"type": "string",
						"minLength": 1,
						"maxLength": 10000
					},
					"validationType": {
						"$ref": "#/components/schemas/ValidationType"
					}
				},
				"required": [
					"evaluationMethodId",
					"sprintId",
					"filePath",
					"validationType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginationResponseDTO_ConsistencyRuleResponseDTO_": {
				"properties": {
					"results": {
						"items": {
							"$ref": "#/components/schemas/ConsistencyRuleResponseDTO"
						},
						"type": "array"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"results",
					"totalPages"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "supportplatform-service",
		"version": "1.0.0",
		"description": "SupportPlatform service for GitHub Project Evaluation Support Platform",
		"license": {
			"name": "CC-BY-4.0"
		},
		"contact": {
			"name": "Lucas Ângelo Oliveira Martins Rocha ",
			"url": "https://lucasangelo.com/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/consistency-rule": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "ConsistencyRule created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsistencyRuleResponseDTO"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"evaluationMethodId": 1,
											"sprintId": 1,
											"standardizedIssueId": 1,
											"description": "ConsistencyRule 1 description",
											"filePath": "/Documentacao/DocumentoDoProjeto.pdf",
											"validationType": "DEFAULT"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new ConsistencyRule.",
				"tags": [
					"consistency-rule"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsistencyRuleCreateDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Found consistency rules",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginationResponseDTO_ConsistencyRuleResponseDTO_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"totalPages": 1,
											"results": [
												{
													"id": 1,
													"evaluationMethodId": 1,
													"sprintId": 1,
													"standardizedIssueId": 1,
													"description": "ConsistencyRule 1 description",
													"filePath": "/Documentacao/DocumentoDoProjeto.pdf",
													"validationType": "DEFAULT"
												},
												{
													"id": 2,
													"evaluationMethodId": 1,
													"sprintId": 1,
													"standardizedIssueId": 1,
													"description": "ConsistencyRule 2 description",
													"filePath": "CITATION.cff",
													"validationType": "CFF"
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all ConsistencyRules with pagination and filter.",
				"tags": [
					"consistency-rule"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": "1",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "evaluationMethodId",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "sprintId",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "standardizedIssueId",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "description",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filePath",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "validationType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ValidationType"
						}
					}
				]
			}
		},
		"/consistency-rule/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "ConsistencyRule updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsistencyRuleResponseDTO"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"evaluationMethodId": 1,
											"sprintId": 1,
											"standardizedIssueId": 1,
											"description": "ConsistencyRule 1 description",
											"filePath": "/Documentacao/DocumentoDoProjeto.pdf",
											"validationType": "DEFAULT"
										}
									}
								}
							}
						}
					}
				},
				"description": "Update an existing ConsistencyRule by id.\nNeed body with all fields.",
				"tags": [
					"consistency-rule"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsistencyRuleUpdateDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Found consistency rule",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsistencyRuleResponseDTO"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 1,
											"evaluationMethodId": 1,
											"sprintId": 1,
											"standardizedIssueId": 1,
											"description": "ConsistencyRule 1 description",
											"filePath": "/Documentacao/DocumentoDoProjeto.pdf",
											"validationType": "DEFAULT"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get an existing ConsistencyRule by id.",
				"tags": [
					"consistency-rule"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}